Android Studio port complete, and reduced required Libraries.
This commit adds the begining of a service framework to do our
background polling.  It is also completing all the details of porting to
Android Studio and the gradle build system, which includes cleaning up
the new binary output files/folders.  Finally, this removes the need for
the SlidingMenu 3rd party library, and replaces it with the
android.v7.appcompat.DrawerLayout pattern from Google's Android support
libaries.